src/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.115.5
flower==2.0.1
h11==0.14.0
humanize==4.11.0
idna==3.10
kombu==5.4.2
prometheus_client==0.21.0
prompt_toolkit==3.0.48
pydantic==2.10.2
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
redis==5.2.0
requests==2.32.3
ruff==0.8.0
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
vine==5.1.0
wcwidth==0.2.13
pytest
httpx
pytest-mock